+++ hashes/src/hash160.rs
@@ -12,7 +12,7 @@ use core::ops::Index;
 use core::slice::SliceIndex;
 use core::str;
 
-use crate::{ripemd160, sha256, Error};
+use crate::{ripemd160, sha256, FromSliceError};
 
 crate::internal_macros::hash_type! {
     160,