+++ hashes/src/sha256t.rs
@@ -8,7 +8,7 @@ use core::ops::Index;
 use core::slice::SliceIndex;
 use core::{cmp, str};
 
-use crate::{sha256, Error};
+use crate::{sha256, FromSliceError};
 
 type HashEngine = sha256::HashEngine;