+++ hashes/src/siphash24.rs
@@ -7,7 +7,7 @@ use core::ops::Index;
 use core::slice::SliceIndex;
 use core::{cmp, mem, ptr, str};
 
-use crate::{Error, Hash as _, HashEngine as _};
+use crate::{FromSliceError, Hash as _, HashEngine as _};
 
 crate::internal_macros::hash_type! {
     64,